+++ user_guide_src/source/changelog.rst
@@ -38,7 +38,8 @@ Release Date: Not Released
 
    -  url_title() will now trim extra dashes from beginning and end.
    -  Added XHTML Basic 1.1 doctype to :doc:`HTML Helper <helpers/html_helper>`.
-   -  Changed humanize to include a second param for the separator.
+   -  Changed humanize() to include a second param for the separator.
+   -  Added date_range() to the :doc:`Date Helper <helpers/html_helper>`.
 
 -  Database